# Meta.from_dict: micro optimization