This is now a bit faster than the custom `from_dict` that we had before (even before https://github.com/iterative/dvc-data/pull/169), as we don't need to call `d.get()` function too many times.

When the dictionary is empty, I see the following results:
```console
Mean +- std dev: [main] 808 ns +- 29 ns -> [s] 581 ns +- 15 ns: 1.39x faster
```

When the dictionary is non-empty, the result is small but positive:
```console
Mean +- std dev: [main2] 862 ns +- 31 ns -> [s2] 837 ns +- 25 ns: 1.03x faster
```